+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'app2'
+rootProject.name = 'PTMate'
