--- src/main/java/com/app2/App2Application.java
@@ -1,20 +1,0 @@
-package com.app2;
-
-import jakarta.annotation.PostConstruct;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-import java.util.TimeZone;
-
-@SpringBootApplication
-public class App2Application {
-
-    public static void main(String[] args) {
-        SpringApplication.run(App2Application.class, args);
-    }
-
-    @PostConstruct
-    public void started() {
-        TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
-    }
-}
