--- src/main/java/com/app2/global/config/SecurityConfig.java
@@ -1,56 +1,0 @@
-package com.app2.global.config;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.annotation.web.configurers.HeadersConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.web.SecurityFilterChain;
-
-@RequiredArgsConstructor
-@Configuration
-@EnableWebSecurity
-@EnableMethodSecurity
-public class SecurityConfig {
-//    private final JwtService jwtService;
-//    private final CustomOAuth2UserService customOAuth2UserService;
-
-    // 아래 URI로 들어오는 요청은 AuthenticationFilter 작동 X
-    private final String[] ALLOWED_URIS = {
-            "/",
-            "/login",
-            "/index.html",
-            "/api/oauth2/**"
-    };
-
-    @Bean
-    public SecurityFilterChain filterChain(HttpSecurity http) throws Exception {
-        http
-                .cors(Customizer.withDefaults())
-                .formLogin(AbstractHttpConfigurer::disable)
-                .httpBasic(AbstractHttpConfigurer::disable)
-                .csrf(AbstractHttpConfigurer::disable)
-                .headers(headers ->
-                        headers.frameOptions(HeadersConfigurer.FrameOptionsConfig::sameOrigin))
-                .sessionManagement(sessionManagement ->
-                        sessionManagement.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-                )
-                .authorizeHttpRequests(request ->
-                        request.requestMatchers(ALLOWED_URIS).permitAll()
-                                .anyRequest().authenticated())
-//                .addFilterAfter(jwtAuthenticationFilter(), LogoutFilter.class)
-        ;
-
-        return http.build();
-    }
-
-//    @Bean
-//    public JwtAuthenticationFilter jwtAuthenticationFilter() {
-//        return new JwtAuthenticationFilter(jwtService, redisService, memberRepository, Collections.singletonList(ALLOWED_URIS));
-//    }
-}
