+++ src/main/java/com/ptmate/PtMateApplication.java
@@ -1,0 +1,21 @@
+package com.ptmate;
+
+import java.util.TimeZone;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+import jakarta.annotation.PostConstruct;
+
+@SpringBootApplication
+public class PtMateApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(PtMateApplication.class, args);
+	}
+
+	@PostConstruct
+	public void started() {
+		TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
+	}
+}
